+++ cmd/js.go
@@ -363,7 +363,7 @@ func openJS() (jetstream.JetStream, func(), error) {
 	if err != nil {
 		return nil, nil, err
 	}
-	nc, settings, err := natsx.Connect(c)
+	nc, settings, err := natsConnect(c)
 	if err != nil {
 		return nil, nil, err
 	}